refactor: 응답 dto ratingAvg 타입 Float에서 Double로 변경

+++ src/main/java/com/sparta/foodorder/domain/store/application/dto/StoreResponseDto.java
@@ -18,7 +18,7 @@ public record StoreResponseDto(
     String phoneNumber,
     Boolean isActive,
     Long ratingCount,
-    Float ratingAvg,
+    Double ratingAvg,
     Long deliveryFee,
     Long minOrderAmount,
     LocalTime opensAt,
